requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests